requirements.txt
requests
python-colored-print
pyfiglet